% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{CONST_LOGIN_TYPES}
\alias{CONST_LOGIN_TYPES}
\title{Login types}
\format{An object of class \code{character} of length 9.}
\usage{
CONST_LOGIN_TYPES
}
\description{
Types of events which count as active users for usage consumption, as a vector
}
\keyword{datasets}
