% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{CONST_LOGIN_TYPES_STR}
\alias{CONST_LOGIN_TYPES_STR}
\title{Login types (string)}
\format{An object of class \code{character} of length 1.}
\usage{
CONST_LOGIN_TYPES_STR
}
\description{
Types of events which count as active users for usage consumption, as a string
}
\keyword{datasets}
