% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_core.R
\name{dwh_connect}
\alias{dwh_connect}
\title{Connect to DWH}
\usage{
dwh_connect(credentials_file = "~/.dwh_credentials", sslmode = "require",
  bigint = "numeric")
}
\arguments{
\item{credentials_file}{The file where the host, database, user, password and port are}

\item{sslmode}{Type of SSL requirement to send to PGSQL's client, Auth0's DWH will only accept SSL connections}

\item{bigint}{The R type that 64-bit integer types should be mapped to, default is numeric (float), which modifies very big numbers, in some scenarios we may need to switch to integer64 to capture it accurately. The possible options are numeric, integer, character and integer64}
}
\description{
Use credentials provided in ini file to connect to DWH, and return the connection pointer
}
\examples{

con = dwh_connect()
tenants = tbl(con, 'tenants')

}
