% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_manipulation.R
\name{dwh_create_view}
\alias{dwh_create_view}
\title{Create view}
\usage{
dwh_create_view(orig_table, view_name, group_read = "DWH_READ")
}
\arguments{
\item{orig_table}{The name of the table to base the view on}

\item{view_name}{The name of the view in Amazon Redshift}

\item{group_read}{The group that should have SELECT permissions on the table}
}
\description{
This function creates a view in Amazon Redshift
}
\examples{

dwh_drop_view('my_view')

}
