% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_core.R
\name{dwh_disconnect}
\alias{dwh_disconnect}
\title{Disconnect from DWH}
\usage{
dwh_disconnect(con)
}
\arguments{
\item{con}{The connection to close}
}
\description{
Close a DWH connection
}
\examples{

con = dwh_connect()
dwh_disconnect(con)

}
