% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_manipulation.R
\name{dwh_drop_view}
\alias{dwh_drop_view}
\title{Drop view}
\usage{
dwh_drop_view(view_name)
}
\arguments{
\item{view_name}{The name of the view in Amazon Redshift}
}
\description{
This function drops an existing view in Amazon Redshift
}
\examples{

dwh_drop_view('my_view')

}
