% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_core.R
\name{dwh_query}
\alias{dwh_query}
\title{DWH Query}
\usage{
dwh_query(query, pcon = NULL, slots = 1, bindValues = NULL)
}
\arguments{
\item{query}{The query text to execute in Amazon Redshift}

\item{pcon}{The connection to run the query on, if unspecified, a temporary connection will be created and closed while running the query}

\item{slots}{The number of slots to run the query on, by default 1}

\item{bindValues}{List of values to bind to the query}
}
\description{
Run specified SQL query in Datawarehouse and return result, allows for specifiying slots and binding values.
}
\examples{

a=dwh_query('select 1', slots=2)
b=dwh_query('select ?, ?', bindValues=c(1,2))

}
