% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_manipulation.R
\name{dwh_replace_inplace_from_sql}
\alias{dwh_replace_inplace_from_sql}
\title{Replace data in Amazon Redshift table in-place}
\usage{
dwh_replace_inplace_from_sql(table_name, query, slots = 1)
}
\arguments{
\item{table_name}{The name of the table in Amazon Redshift}

\item{query}{The query to execute to create the table in Amazon Redshift}

\item{slots}{The number of slots to use}
}
\description{
This function replaces a table from an SQL query without dropping/creating the destination table, making it suitable for tables which may be used while the process is running, but does not allow for structure changes without altering the original table.
}
\examples{

dwh_replace_inplace_from_sql('test_table', 'select 1')

}
