% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_manipulation.R
\name{dwh_replace_table_from_sql}
\alias{dwh_replace_table_from_sql}
\title{Replace DWH Table from SQL}
\usage{
dwh_replace_table_from_sql(table_name, query, group_read = "DWH_READ",
  group_all = "DWH", slots = 1, pcon = NULL)
}
\arguments{
\item{table_name}{The name of the table in Amazon Redshift}

\item{query}{The query to execute to create the table in Amazon Redshift}

\item{group_read}{The group that will have read access to the table}

\item{group_all}{The group that will have all access to the table}

\item{slots}{The number of slots to use}

\item{pcon}{Optionally, use an existing connection, if not, will start a temporary connection to use}
}
\description{
Replace data using an SQL query. The existing table will be dropped if it exists.
}
\examples{

dwh_replace_table_from_sql('test_table', 'select 1')

}
