% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_manipulation.R
\name{dwh_replace_table_from_temp_table}
\alias{dwh_replace_table_from_temp_table}
\title{Replace DWH Table from temporary table}
\usage{
dwh_replace_table_from_temp_table(con, table_name, temp_table, slots = 1,
  transaction = T)
}
\arguments{
\item{con}{Connection where to use the temporary table}

\item{table_name}{The name of the table in Amazon Redshift to copy into}

\item{temp_table}{The name of the temporary table in Amazon Redshift}

\item{slots}{The number of slots to use}

\item{transaction}{Use begin/commit commands, by default yes}
}
\description{
Replace data using another temporary table. The existing table will be truncated.
}
\examples{

con = dwh_connect()
# [...] generate temp table
dwh_replace_table_from_temp_table(con, 'test_table', 'temp_table')

}
