% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_core.R
\name{dwh_set_execution_slots}
\alias{dwh_set_execution_slots}
\title{Set execution slots}
\usage{
dwh_set_execution_slots(con, slots)
}
\arguments{
\item{con}{The connection to set slots on}

\item{slots}{The number of slots to use}
}
\description{
To add processing power to queries, you can add query slots in a specific session, take into account this will slow other queries running in the system so it has to be done very sparingly for critical time-sensitive queries in the critical path. Read more at https://docs.aws.amazon.com/redshift/latest/dg/c_workload_mngmt_classification.html
}
\examples{

con = dwh_connect()
dwh_set_execution_slots(con, 2)

}
