% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_metadata.R
\name{dwh_table_schema}
\alias{dwh_table_schema}
\title{Get table schema from DWH}
\usage{
dwh_table_schema(table_names, schema = "public")
}
\arguments{
\item{table_names}{Name of the tables to get schema}

\item{schema}{Name of the schema where the table is, defaults to public}
}
\description{
Obtains columns, primary keys, foreign key relationships not null and column order  from Amazon Redshift
}
\examples{

schema = dwh_table_schema(c('tenants', 'users'))

}
