% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scd_etl.R
\name{dwh_update_scd_type1_in_database}
\alias{dwh_update_scd_type1_in_database}
\title{SCD Type 1 in-database}
\usage{
dwh_update_scd_type1_in_database(con, dim_schema, dim_name, new_ref,
  surrogate_key, natural_keys)
}
\arguments{
\item{con}{Connection to use}

\item{dim_schema}{Schema of the dimension to update}

\item{dim_name}{Table name to update}

\item{new_ref}{The data you want to insert into the dimension, it has to be a tbl reference in the same connection as con}

\item{surrogate_key}{The surrogate key you want to use, it has to be an integer column, and it has to be the first column in the table}

\item{natural_keys}{The natural keys to use for joining the old data with the new data}
}
\description{
Generate changes for a Slowly changing dimension type 1 (Overwrite values) using an existing DWH table and a tibble database reference
}
\examples{
\dontrun{

con = dwh_connect()
cur_dim = tbl(con, in_schema('prod', 'dim_my_dimension'))
new_data = mutate(cur_dim, bla=1) \%>\% compute()
dwh_update_scd_type1_in_database(con, 'prod', 'my_dimension',
new_data, dim_key, c('business_key', 'business_key2'))

}

}
