% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scd_etl.R
\name{dwh_update_scd_type2}
\alias{dwh_update_scd_type2}
\title{SCD Type 2}
\usage{
dwh_update_scd_type2(current_dim, new_data, surrogate_key, natural_keys,
  scd_date = Sys.Date())
}
\arguments{
\item{current_dim}{Current table, if a reference to a table in DWH, will download the entire table into R}

\item{new_data}{The data you want to insert into the dimension}

\item{surrogate_key}{The surrogate key you want to use, it has to be an integer column, and it has to be the first column in the table}

\item{natural_keys}{The natural keys to use for joining the old data with the new data}

\item{scd_date}{The date to insert, defaults to the current date}
}
\description{
Generate changes for a Slowly changing dimension type 2 (Add new rows)
}
\examples{
\dontrun{

cur_dim = dwh_query('select * from prod.dim_my_dimension')
new_data = mutate(cur_dim, bla=1)
changes = dwh_update_scd_type2(cur_dim, new_data, dim_key, c('business_key', 'business_key2'))

}
}
