% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_manipulation.R
\name{dwh_upsert_table_from_temp_table}
\alias{dwh_upsert_table_from_temp_table}
\title{Upsert DWH Table from temporary table}
\usage{
dwh_upsert_table_from_temp_table(con, table_name, temp_table, keys, slots = 1)
}
\arguments{
\item{con}{Connection where to use the temporary table}

\item{table_name}{The name of the table in Amazon Redshift to copy into}

\item{temp_table}{The name of the temporary table in Amazon Redshift}

\item{keys}{The keys to upsert}

\item{slots}{The number of slots to use}
}
\description{
Upsert data using another temporary table.
}
\examples{

con = dwh_connect()
# [...] generate temp table
dwh_upsert_table_from_temp_table(con, 'test_table', 'temp_table', c('id', 'id2'))

}
