% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metadata.R
\name{get_metadata_value}
\alias{get_metadata_value}
\title{Get metadata value}
\usage{
get_metadata_value(metadata_key, source = "dwh")
}
\arguments{
\item{metadata_key}{The key of the metadata value, e.g. "last_processed_fit_score" could be the last processed time of the fit score}

\item{source}{if `dwh`, it'll use the DWH, if you specify anything else, it'll use it as a filename to use as a config database}
}
\description{
Read metadata value from DWH, will autodetect field type and pull from appropiate table
}
\examples{
\dontrun{

end_date = get_metadata_value('last_processed_fit_score')
print(end_date)

}
}
