% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_color_auth0_discrete}
\alias{scale_color_auth0_discrete}
\alias{scale_fill_auth0_discrete}
\alias{scale_fill_auth0_gradient}
\alias{scale_color_auth0_gradient}
\alias{scale_colour_auth0_discrete}
\alias{scale_colour_auth0_gradient}
\title{Color scales defined by Auth0's design team}
\usage{
scale_color_auth0_discrete(..., palette = "default", direction = 1,
  aesthetics = "colour")

scale_fill_auth0_discrete(..., palette = "default", direction = 1,
  aesthetics = "fill")

scale_fill_auth0_gradient(..., midpoint = 0, space = "Lab",
  na.value = "grey50", guide = "colourbar", aesthetics = "fill")

scale_color_auth0_gradient(..., midpoint = 0, space = "Lab",
  na.value = "grey50", guide = "colourbar", aesthetics = "colour")
}
\arguments{
\item{...}{Other arguments passed on to [discrete_scale()] to control name,
limits, breaks, labels and so forth.}

\item{palette}{The palette to use, if left empty will use the default one, you can also choose the sequential one if the value is ordinal, or colorful to use many different colors.}

\item{direction}{The direction to use, if -1 will return the colors in opposite order}

\item{aesthetics}{The aesthetic to apply this scale on}

\item{midpoint}{The mid value on gradients, should always set this to the mean or median of the value to colour.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" - other values are deprecated.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use "colourbar" for continuous colour bar, or "legend" for discrete colour legend.}
}
\description{
The `auth0` scales provides color palettes provided by Auth0's design team. They are meant to be used in combination with the auth0_theme in ggplot2
}
\details{
The `auth0` scales were carefully designed and tested on discrete data.
}
\section{Palettes}{

The following palettes are available for use with these scales:
\describe{
  \item{default}{Default palette alternating between colors ranging from yellow to orange}
  \item{sequential}{Colors ordered from lighter to darker, meant for ordinal variables}
  \item{colorful}{Colorful alternative respecting Auth0's style}
}
}

\examples{

ggplot(diamonds, aes(clarity, fill = cut)) +
geom_bar() +
theme_auth0() +
scale_fill_auth0_discrete(palette='colorful')

ggplot(diamonds, aes(x= cut, y = depth, color = cut)) +
geom_violin(fill = '#fffffa') +
theme_auth0() +
scale_color_auth0_discrete()

ggplot(diamonds, aes(x= depth, y = price, color = price)) +
geom_point() + theme_auth0() + scale_color_auth0_gradient(midpoint=10000)

}
