% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{SCD_FUTURE_DATE}
\alias{SCD_FUTURE_DATE}
\title{SCD Future date}
\format{An object of class \code{Date} of length 1.}
\usage{
SCD_FUTURE_DATE
}
\description{
Default time for future values on SCDs with dates
}
\keyword{datasets}
