% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{SCD_START_DATE}
\alias{SCD_START_DATE}
\title{SCD Start date}
\format{An object of class \code{Date} of length 1.}
\usage{
SCD_START_DATE
}
\description{
Default time for blank values on SCDs with dates
}
\keyword{datasets}
