% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metadata.R
\name{set_metadata_value}
\alias{set_metadata_value}
\title{Set metadata value}
\usage{
set_metadata_value(metadata_key, value, source = "dwh")
}
\arguments{
\item{metadata_key}{The key of the metadata value, e.g. "last_processed_fit_score" could be the last processed time of the fit score}

\item{value}{The data you want to save under this key}

\item{source}{if `dwh`, it'll use the DWH, if you specify anything else, it'll use it as a filename to use as a config database}
}
\description{
Set the value of a key to the requested value
}
\examples{
\dontrun{

set_metadata_value('my_cool_value', 'hello')

}
}
