% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-dbplyr-redshift.R
\name{sql_semi_join.PqConnection}
\alias{sql_semi_join.PqConnection}
\title{SQL Semi join to replace dbplyr's one}
\usage{
\method{sql_semi_join}{PqConnection}(con, x, y, anti = FALSE, by = NULL,
  ...)
}
\arguments{
\item{con}{PGSQL Connection}

\item{x}{Left table}

\item{y}{Right table}

\item{anti}{Anti join?}

\item{by}{By which variable/s}

\item{...}{Other params to pass to underlying methods}
}
\description{
SQL Semi join to replace dbplyr's one
}
