% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_auth0}
\alias{theme_auth0}
\title{Theme defined by Auth0's design team}
\usage{
theme_auth0()
}
\description{
The `auth0` theme provides a carefully designed theme with help from Auth0's design team. They are meant to be used in combination the auth0 palettes
}
\examples{

ggplot(diamonds, aes(clarity, fill = cut)) +
geom_bar() +
theme_auth0() +
scale_fill_auth0_discrete(palette='colorful')

}
